Paginação base

+++ WeCare.Domain/Models/ParticipationCertificate.cs
@@ -5,7 +5,7 @@ public class ParticipationCertificate
     public long Id { get; set; }
     public DateTime CreationDate { get; set; }
 
-    public int CandidateId { get; set; }
+    public long CandidateId { get; set; }
     public Candidate? Candidate { get; set; }
 
     public long VolunteerOpportunityId { get; set; }
